Add an example of using hash codes using c++ HashTable.cpp
I got this code from https://www.digitalocean.com/community/tutorials/hash-table-in-c-plus-plus
This is only the beginning of the article.  I added the who am I code at
the start.

Modified the Makefile to compike the HashTable.cpp

Add the code.

Hmm the pre-commit hook  prevented it from adding because of a blank line at end.